Use abbot-cli to manage the 404-finder skill
